Add files via upload
Americans might tend to inflate the influence of the United States in the history of the spread of English. Before the World Wars, particularly WWII, the US was a bit player on the world stage. The folks with the big international footprint were England, France, Spain, and Portugal. 

Look around the world and you can see the linguistic legacy of the colonial era stamped all over the globe. India, Australia, Kenya, and Guyana were all under England's kind protection, and English is the primary or secondary language in all of them. France made its biggest play in West Africa, and you'd better know French if you want to get around in that part of the world without knowing a local language like Wolof or Pulaar. Spain conquered most of the Americas south of the present-day US, so your best language to get around most countries in the Western Hemisphere is Spanish. Portugal had Brazil and some big territories in Africa - learn Portuguese if you plan to attend the next World Cup or Summer Ol